services/email/requirements.txt
pika==1.3.2
pika-stubs==0.1.3
python-dotenv==1.0.1
secure-smtplib==0.1.1